Added the context menu we talked about to the posts table view.
I *think* I did this in the "right" way, if there is a right way. I
made a menu in IB, hooked it up to the "menu" outlet of the table view,
and then set up each menu item to send an IBAction to the document
class. The document class uses the table view's clickedRow property to
figure out what post to edit/preview.