Improved the file watching system for the app.
I made it so that if the preview server isn't running at all, the app
will still watch the Posts directory for any changes. Whenever it
detects a change in the Posts directory, it will call parsePosts on the
site object, meaning that the UI will get updated but no files will get
written.