Fixed a bug where saving a post file wouldn't trigger the auto-refresh method.
I kept blaming this on Lion's autosave/versions system. I thought, "It works just fine for images and stuff! It must be something about documents and whatnot." Wrong. As it turns out, the reason that it worked for other types of files but not posts is because *the posts directory wasn't added to the list of paths to watch.* In other words, I'm going to go cry now.