Added basic AppleScript support to Tribo.
This is far from complete, but it has reached a point where it can be
used to retrieve structured information from the app. The largest
omission is the ability to edit…anything. Every property is currently
read-only. I'm not sure how I want to handle editing, because no
editing is supported through the app yet at all, and support for
undoing edits would be important.

Because this is incomplete, it does not fully address #8.