Enabled sandboxing.
This commit includes a .p12 file which contains a certificate. The
certificate can be used to sign the app for sandbox testing. It is a
self-signed certificate, and thus cannot be used for iCloud testing or
Mac App Store submission.

I will write up instructions for how to install the certificate soon
and put them in the Readme.