Reworked quite a bit of the app's UI.
The original bottom bar is gone (meaning this closes #28). It has been
replaced by an action menu triggered by the accessory view in the
window, along with a dynamic status bar which appears in the bottom of
the window when the local preview is running. Hopefully the status bar
will also be used to show publishing progress in a future commit.

This commit includes artwork and color schemes from Wil Nichols
(http://wilnichols.com/). Thanks Wil!