Began implementing syntax highlighting for the post editor.
I got as far as header sizing, but it’s really just a proof-of-concept at this point. I need to implement a better way of cascading styles onto parts of text, and generalize the storage of the styles so that I can pull them out of a JSON or Plist file at runtime.