Fixed a long-standing drawing bug with the tab view.
Ever since I wrote it, the tab view has had a problem with the drawing getting out-of-whack after the window resizes. After quite a bit of debugging, it seems to have something to do with varying rounding behaviors for view frames during autoresizing. Because of that, the frames of the individual tabs themselves where what you might call “unreliable”. The frames they were reporting didn’t line up with the frames I was calculating. To work around this whole issue, I started the -drawRect: method by pre-calculating all the tab frames, and then using my calculations for drawing. Now everything lines up, and the view resizes correctly. Hooray for math!