Added live reload support to the CLI.
This was pretty tricky. The two LiveReload files had to be placed inside the Info.plist file, which is then embedded into the __TEXT section of the CLI. I can then access the keys inside via NSBundle methods. It’s a bit of a hack, but it keeps the CLI self-contained and doesn’t duplicate any data.